pyproject.toml
[project]
name = "nighthawk"
version = "0.1.0"
description = "An experimental Python library that embeds Natural blocks inside Python functions and executes them using an LLM."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "headson>=0.16.1",
  "pydantic>=2",
  "pydantic-ai-slim>=1.59",
  "pyyaml>=6",
  "tiktoken>=0.12",
]

[project.optional-dependencies]
openai = [
  "pydantic-ai-slim[openai]>=1.59",
]
vertexai = [
  "pydantic-ai-slim[google,vertexai]>=1.59",
]
claude-code = [
  "claude-agent-sdk>=0.1",
]
codex = [
  "mcp>=1.26",
]

[dependency-groups]
dev = [
  "logfire>=4",
  "pytest>=9",
  "pytest-asyncio>=1.3.0",
  "ruff>=0.15",
]
docs = [
  "mkdocs>=1.6",
  "mkdocs-material>=9",
  "mkdocstrings[python]>=0.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nighthawk"]

[tool.ruff]
target-version = "py313"
line-length = 150

[tool.ruff.lint]
extend-select = ["I", "UP", "B", "SIM", "N"]

[tool.ruff.lint.isort]
known-first-party = ["nighthawk"]
